Fixing rebase after moving saved recordings to history
